sinefill clone in waveshap.h

+++ dsp/verb.hpp
@@ -12,9 +12,9 @@ namespace dspheaders{
       float fb;
       // __m128 fb;
       int writeptr = 0;
-      float v[4] {0,    0,     0,    0    };
-      float k[4] {0.8f, 0.35f, 0.5f, 0.12f};
-      int   i[4] {17,   19,    29,   47   };
+      float v[4]   { 0,    0,     0,    0     };
+      float k[4]   { 0.8f, 0.35f, 0.5f, 0.12f };
+      float   i[4] { 17.f, 19.f,  29.f, 47.f  };
 
       // void write();
       // float read(int pos);
